+++ app/clients/internet/__init__.py
@@ -1,0 +1,4 @@
+from ._braveinternetclient import BraveInternetClient
+from ._duckduckgointernetclient import DuckDuckGoInternetClient
+
+__all__ = ["BraveInternetClient", "DuckDuckGoInternetClient"]