+++ app/utils/variables.py
@@ -1,8 +1,15 @@
-AUDIO_MODEL_TYPE = "automatic-speech-recognition"
+DEFAULT_TIMEOUT = 120
+
 INTERNET_COLLECTION_DISPLAY_ID = "internet"
+
+HYBRID_SEARCH_TYPE = "hybrid"
+LEXICAL_SEARCH_TYPE = "lexical"
+SEMANTIC_SEARCH_TYPE = "semantic"
+
 PUBLIC_COLLECTION_TYPE = "public"
 PRIVATE_COLLECTION_TYPE = "private"
 
+AUDIO_MODEL_TYPE = "automatic-speech-recognition"
 EMBEDDINGS_MODEL_TYPE = "text-embeddings-inference"
 LANGUAGE_MODEL_TYPE = "text-generation"
 
@@ -13,137 +20,10 @@ PDF_TYPE = "application/pdf"
 JSON_TYPE = "application/json"
 TXT_TYPE = "text/plain"
 HTML_TYPE = "text/html"
-
 # @TODO : add DOCX_TYPE (application/vnd.openxmlformats-officedocument.wordprocessingml.document)
-ROLE_LEVEL_0 = 0
-ROLE_LEVEL_1 = 1
-ROLE_LEVEL_2 = 2
-USER_ROLE = "user"
 
-INTERNET_DUCKDUCKGO_TYPE = "duckduckgo"
-INTERNET_BRAVE_TYPE = "brave"
-
-SEARCH_ELASTIC_TYPE = "elastic"
-SEARCH_QDRANT_TYPE = "qdrant"
-
-
-SUPPORTED_LANGUAGES = {
-    "afrikaans": "af",
-    "albanian": "sq",
-    "amharic": "am",
-    "arabic": "ar",
-    "armenian": "hy",
-    "assamese": "as",
-    "azerbaijani": "az",
-    "bashkir": "ba",
-    "basque": "eu",
-    "belarusian": "be",
-    "bengali": "bn",
-    "bosnian": "bs",
-    "breton": "br",
-    "bulgarian": "bg",
-    "burmese": "my",
-    "cantonese": "yue",
-    "castilian": "es",
-    "catalan": "ca",
-    "chinese": "zh",
-    "croatian": "hr",
-    "czech": "cs",
-    "danish": "da",
-    "dutch": "nl",
-    "english": "en",
-    "estonian": "et",
-    "faroese": "fo",
-    "finnish": "fi",
-    "flemish": "nl",
-    "french": "fr",
-    "galician": "gl",
-    "georgian": "ka",
-    "german": "de",
-    "greek": "el",
-    "gujarati": "gu",
-    "haitian": "ht",
-    "haitian creole": "ht",
-    "hausa": "ha",
-    "hawaiian": "haw",
-    "hebrew": "he",
-    "hindi": "hi",
-    "hungarian": "hu",
-    "icelandic": "is",
-    "indonesian": "id",
-    "italian": "it",
-    "japanese": "ja",
-    "javanese": "jw",
-    "kannada": "kn",
-    "kazakh": "kk",
-    "khmer": "km",
-    "korean": "ko",
-    "lao": "lo",
-    "latin": "la",
-    "latvian": "lv",
-    "letzeburgesch": "lb",
-    "lingala": "ln",
-    "lithuanian": "lt",
-    "luxembourgish": "lb",
-    "macedonian": "mk",
-    "malagasy": "mg",
-    "malay": "ms",
-    "malayalam": "ml",
-    "maltese": "mt",
-    "mandarin": "zh",
-    "maori": "mi",
-    "marathi": "mr",
-    "moldavian": "ro",
-    "moldovan": "ro",
-    "mongolian": "mn",
-    "myanmar": "my",
-    "nepali": "ne",
-    "norwegian": "no",
-    "nynorsk": "nn",
-    "occitan": "oc",
-    "panjabi": "pa",
-    "pashto": "ps",
-    "persian": "fa",
-    "polish": "pl",
-    "portuguese": "pt",
-    "punjabi": "pa",
-    "pushto": "ps",
-    "romanian": "ro",
-    "russian": "ru",
-    "sanskrit": "sa",
-    "serbian": "sr",
-    "shona": "sn",
-    "sindhi": "sd",
-    "sinhala": "si",
-    "sinhalese": "si",
-    "slovak": "sk",
-    "slovenian": "sl",
-    "somali": "so",
-    "spanish": "es",
-    "sundanese": "su",
-    "swahili": "sw",
-    "swedish": "sv",
-    "tagalog": "tl",
-    "tajik": "tg",
-    "tamil": "ta",
-    "tatar": "tt",
-    "telugu": "te",
-    "thai": "th",
-    "tibetan": "bo",
-    "turkish": "tr",
-    "turkmen": "tk",
-    "ukrainian": "uk",
-    "urdu": "ur",
-    "uzbek": "uz",
-    "valencian": "ca",
-    "vietnamese": "vi",
-    "welsh": "cy",
-    "yiddish": "yi",
-    "yoruba": "yo",
-}
-
-HYBRID_SEARCH_TYPE = "hybrid"
-LEXICAL_SEARCH_TYPE = "lexical"
-SEMANTIC_SEARCH_TYPE = "semantic"
-
-DEFAULT_TIMEOUT = 120
+# Clients
+SEARCH_CLIENT_ELASTIC_TYPE = "elastic"
+SEARCH_CLIENT_QDRANT_TYPE = "qdrant"
+INTERNET_CLIENT_BRAVE_TYPE = "brave"
+INTERNET_CLIENT_DUCKDUCKGO_TYPE = "duckduckgo"